further split routing_renormalize.py to 3 scripts
